Fix for MISO-193: popup windows occasionally don't pop up
jQueryUI dialog creation had been wrapped in an anonymous jQuery
function and then opened, but the anonymous jQuery function was not
immediately executed (and would sometimes not execute, causing the
reported behavior).

Since the jQueryUI dialog creation will only be invoked when the
surrounding JavaScript function is called, the anonymous jQuery function
and the "autoOpen: false" flag are not required.