GLT-431 Made miso-dto available in the miso-service layer. This allows dtos to be passed from miso-web to miso-service. Once in miso-service the dtos can be converted to model objects and items referred to by id can be retrieved from the db.
GLT-431 Changes to create new Analyte sample. Will also create an
Identity sample if no parent is provided.
- Updated transactions so that nested calls would all use the same
  transaction.
- Modified authorization to allow users to create samples (not just
  admin).
- Added 'to' service methods to build model objects from DTOs.
- Check to ensure that parent/child relationship is correct.

GLT-431 Turned KitDescriptorDto nested in SampleAdditionalInfoDto into
an id, prepKitId. This works with the sample form since the user will
choose the kit from a drop down list and and provide the form with an
id.

GLT-431 Removed an incorrect check for parent. Corrected the conversion
of SampleTissue to SampleTissueDto within the SampleDto.

GLT-431 Corrected id name of SampleTissue to match name in db and be
consistent with other tables.

GLT-431 Check if identity external name already exists. Use this
identity if it does. If not, generate a new internal Identity name.

GLT-431 Modified checks to return IllegalArgument exceptions instead of
NPEs. Propogated transactions for identity lookup and identity number
generation. Changed REST name from 'fred' to 'tree' (as in hierarchy).

GLT-431 The miso-dto module has core as a dependency in order to access
the model objects.

GLT-431 Implemented creation of a SampleTissue. Reported error when
lookup of id returns null.