Delete plates
This happening because plates really don't do what they are supposed to. The
idea of a plate is that samples would be loaded into a plate and a robot would
do the subsequent procedures to create libraries, dilutions, and pools in that
plate and then the whole plate could be kept in the freezer. This makes plates
really weird objects since they can be used semi-interchangeably with any of
those items as a collection of the them (excluding the entities contained with
in that aren't. The problem is that either all the corresponding data (_e.g._,
library indices) have to be stored in real libraries or as part of the plate.
This model of plate doesn't track all of those things and if all that data is
being tracked using the normal models, then the plate is just a weird kind of
immutable box that is also poolable.

The preferred strategy is to make plates into a workflow that collects all the
right data and, in one step, populates a full hierarchy of model objects and
puts them in a box. This eliminates the special case handling of plates in may
places and whittles down pools to containing only emPCR dilutions and library
dilutions, paving the way for future elimination of the `Poolable` type.