Refactor and fix problems with inferring tissue for tissue processing
This fixes a bug where tissue class was not inferred for tissue processing
samples created directly. It also shares that logic for all sample class
inference. Additionally, it limits the list of available option so that new
samples cannot be created from a class with no inference path to an identity.

JIRA: GLT-1922