[GLT-2916] Patch 'No Scientific Name' bug (#2009)
Putting periods in field names should cause Handsontable to create nested objects and fields. In rare circumstances, it fails to do this, creating a situation where on save there is an extra field flat["sample.scientificName"] which contains a string but flat.sample.scientificName is null. Detect when this happens and force the value back to the correct place to enable saving. Does not fix the underlying issue, which may lie in Handsontable.

Correct a typo in an unrelated function.