Make Library validator not do lookups.
Now the Library validator operates only on range-based regex, rather
than performing database lookups and building regex from that. This is
how the Sample validator works, and is a lot less strict than the
previous version. It still has at least one undocumented assumption built
in that I will file tickets for documenting.

Is this preferable?